third_party/WebKit/LayoutTests/external/wpt/eventsource/shared-worker/eventsource-constructor-non-same-origin-expected.txt

This is a testharness.js-based test.
FAIL shared worker - EventSource: constructor (act as if there is a network error) (http://example.not) assert_equals: source.readyState expected 2 but got 0
FAIL shared worker - EventSource: constructor (act as if there is a network error) (https://example.not/test) assert_equals: source.readyState expected 2 but got 0
PASS shared worker - EventSource: constructor (act as if there is a network error) (ftp://example.not)
PASS shared worker - EventSource: constructor (act as if there is a network error) (about:blank)
PASS shared worker - EventSource: constructor (act as if there is a network error) (mailto:whatwg@awesome.example)
PASS shared worker - EventSource: constructor (act as if there is a network error) (javascript:alert('FAIL'))
Harness: the test ran to completion.
